// frontend/src/pages/Quiz/Constructor/Constructor.types.ts
export type QuestionType =
  | "single"
  | "multi"
  | "binary"
  | "dropdown"
  | "text"
  | "textarea"
  | "date"
  | "welcome"
  | "personal"
  | "completion"
  | "thankyou";

export interface BaseItem {
  type: QuestionType;
  title: string;
  description: string;
}

export interface QuestionItem extends BaseItem {
  options?: string[];
}

export interface ConstructorState {
  items: QuestionItem[];
}